message-image/requirements.txt
Flask==2.0.3
SQLAlchemy
Flask-Migrate
psycopg2-binary==2.9.1
Werkzeug==2.0.3


